auth_api/requirements.txt
SQLAlchemy==2.0.9
psycopg2-binary==2.9.6
alembic==1.10.3
python-dotenv==1.0.0
flask==2.2.3
pydantic==1.10.7
flask-jwt-extended==4.4.4
flask-restful==0.3.9
gevent==22.10.2
flask-wtf==1.1.1
cryptography==40.0.1
flask-sqlalchemy==3.0.3
redis==4.5.4